Primitives deserialization features
