+++ src/compiler/serialization/mod.ts
@@ -4,6 +4,7 @@ export * from './ucs-function.js';
 export * from './ucs-generator.js';
 export * from './ucs-lib.js';
 export * from './ucs-support-bigint.js';
+export * from './ucs-support-boolean.js';
 export * from './ucs-support-defaults.js';
 export * from './ucs-support-list.js';
 export * from './ucs-support-map.js';
