+++ src/compiler/serialization/ucs-support-bigint.ts
@@ -7,10 +7,10 @@ import { UcsCompiler } from './ucs-compiler.js';
 
 export function ucsSupportBigInt(
   compiler: UcsCompiler,
-  target: UcBigInt.Schema | UcDataType<UcBigInt>,
-): UccConfig<UcBigInt.Variant> {
+  target: UcBigInt.Schema | UcDataType<UcBigInt> = BigInt,
+): UccConfig<UcBigInt.Variant | void> {
   return {
-    configure({ string = 'parse', number = 'parse' }) {
+    configure({ string = 'parse', number = 'parse' } = {}) {
       compiler.useUcsGenerator(target, (_fn, _schema, { writer, value }) => code => {
         if (string === 'serialize') {
           const ucsApostrophe = UC_MODULE_SERIALIZER.import('UCS_APOSTROPHE');
